4 créer un service web

+++ cicd-app/src/city/city.controller.ts
@@ -1,15 +1,27 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  HttpCode,
+  HttpStatus,
+  HttpException,
+} from '@nestjs/common';
 import { CityService } from './city.service';
 import { CreateCityDto } from './dto/create-city.dto';
 import { UpdateCityDto } from './dto/update-city.dto';
+import { City } from '@prisma/client';
 
 @Controller('city')
 export class CityController {
   constructor(private readonly cityService: CityService) {}
 
   @Post()
-  create(@Body() createCityDto: CreateCityDto) {
-    return this.cityService.create(createCityDto);
+  create(@Body() body: CreateCityDto) {
+    return this.cityService.create(body);
   }
 
   @Get()
@@ -23,12 +35,20 @@ export class CityController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateCityDto: UpdateCityDto) {
+  async update(@Param('id') id: string, @Body() updateCityDto: UpdateCityDto) {
+    const city: City | null = await this.cityService.findOne(+id);
+    if (!city) {
+      throw new HttpException('City not found', HttpStatus.NOT_FOUND);
+    }
     return this.cityService.update(+id, updateCityDto);
   }
-
+  
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+  const city: City | null = await this.cityService.findOne(+id);
+  if (!city) {
+    throw new HttpException('City not found', HttpStatus.NOT_FOUND);
+  }
     return this.cityService.remove(+id);
   }
 }
